Updating CLI workflow (#692)
Made some minor edits to the doc to clarify that when you do "appcenter login" it opens a browser and generates a new API token for you. I am not sure why we call it an authentication token here, but later tell a user to create their own API token.