Add further information regarding shared schemes.
For most projects, the chance that the `xcshareddata` is in `.gitignore` is high. Coupled with the suggestion that `App Center automatically detects the projects, workspaces, and shared schemes in your branch.`  (cf. https://docs.microsoft.com/en-us/appcenter/build/ios/first-build#31-projectworkspace-and-scheme), this led me to believe that I could simply export the `.xcscheme` and put it anywhere in the project. The addition of this new paragraph is intended to clarify this. I will also propose a change on the other page that is linked here.