Update React Native build script examples
I've added a few extra examples to the custom build script documentation
for React Native. I think these examples are a more realistic
representation of the challenges faced by JavaScript and React Native
projects running in CI environments. This includes authenticating with a
private npm registry and creating secret files that are not in version
control.