Dashboard overview
Deleted image which was outdated & inconsistent with App Center UI, due to: 
- Removal of MBaaS
- Display of CodePush as separate beacon
- No replacement image provided

Also removed references to MBaaS services, and added in-line links to docs subsections which serves to highlight the different options and point readers to where they can learn more about them. I think that the approach of in-line links is a better solution then the static image, though I can add an updated image back if desired & provided with a replacement.