Remove wrong parameter from symbol upload call
This was added in #1956 under the pretense of making sure we send a correct content length to API management.
This request will not hit API management at this time. It will be to a SAS URL in the Azure storage account to upload the files to.

Additionally, this breaks the upload command, as per curl spec, specifying `-d`:
> This option overrides -F, --form and -I, --head and -T, --upload-file.